Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/character_solver/point_triangle_vertex_error_function.cpp
@@ -703,7 +703,7 @@ double PointTriangleVertexErrorFunctionT<T>::getJacobian(
           jacobian.block(3 * iCons, 0, 3, modelParameters.size()),
           residual.middleRows(3 * iCons, 3));
     }
-    usedRows = 3 * constraints_.size();
+    usedRows = static_cast<int>(3 * constraints_.size());
   } else {
     MT_PROFILE_EVENT("VertexErrorFunction - normal jacobians");
     T sourceNormalWeight;
@@ -722,7 +722,7 @@ double PointTriangleVertexErrorFunctionT<T>::getJacobian(
           residual(iCons));
     }
 
-    usedRows = constraints_.size();
+    usedRows = static_cast<int>(constraints_.size());
   }
 
   return error;
